move status creation to controller
