+++ controller/status_test.go
@@ -313,3 +313,158 @@ func Test_Status_AggregateStatus(t *testing.T) {
 		}
 	}
 }
+
+func Test_Common_CreateStatus(t *testing.T) {
+	RegisterTestingT(t)
+
+	type input struct {
+		Group  string
+		USL    UnitStatusList
+		Expand bool
+	}
+	type testCase struct {
+		Comment  string
+		Input    input
+		Expected []string
+	}
+
+	testCases := []testCase{
+		// 3 slices of 1 group with 2 units
+		testCase{
+			Comment: "3 slices of 1 group with 2 units",
+			Input: input{
+				Group: "example",
+				USL: UnitStatusList{
+					givenUnitStatus("example-foo@1.service", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "loaded"),
+					givenUnitStatus("example-bar@1.service", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "loaded"),
+					givenUnitStatus("example-foo@2.service", "2", "172.17.8.102", "9ebb53b04b0d46fb94b4fd1b3f562d2b", "loaded", "loaded"),
+					givenUnitStatus("example-bar@2.service", "2", "172.17.8.102", "9ebb53b04b0d46fb94b4fd1b3f562d2b", "loaded", "loaded"),
+					givenUnitStatus("example-foo@3.service", "3", "172.17.8.103", "e3cb5f13a9164ba5b7eff6c920475e61", "loaded", "loaded"),
+					givenUnitStatus("example-bar@3.service", "3", "172.17.8.103", "e3cb5f13a9164ba5b7eff6c920475e61", "loaded", "loaded"),
+				},
+				Expand: false,
+			},
+			Expected: []string{
+				"Group | Units | FDState | FCState | SAState | IP | Machine",
+				"",
+				"example@1 | * | loaded | loaded | inactive | 172.17.8.101 | 505e0d7802d7439a924c269b76f34b5f",
+				"example@2 | * | loaded | loaded | inactive | 172.17.8.102 | 9ebb53b04b0d46fb94b4fd1b3f562d2b",
+				"example@3 | * | loaded | loaded | inactive | 172.17.8.103 | e3cb5f13a9164ba5b7eff6c920475e61",
+				"",
+			},
+		},
+		// 1 slice of 1 group with 2 units
+		testCase{
+			Comment: "1 slice of 1 group with 2 units",
+			Input: input{
+				Group: "example",
+				USL: UnitStatusList{
+					givenUnitStatus("example-foo", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "loaded"),
+					givenUnitStatus("example-bar", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "loaded"),
+				},
+				Expand: false,
+			},
+			Expected: []string{
+				"Group | Units | FDState | FCState | SAState | IP | Machine",
+				"",
+				"example@1 | * | loaded | loaded | inactive | 172.17.8.101 | 505e0d7802d7439a924c269b76f34b5f",
+				"",
+			},
+		},
+		// 3 slices of 1 group with 2 units - verbose
+		testCase{
+			Comment: "3 slices of 1 group with 2 units - verbose",
+			Input: input{
+				Group: "example",
+				USL: UnitStatusList{
+					givenUnitStatus("example-foo@1.service", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "loaded"),
+					givenUnitStatus("example-bar@1.service", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "loaded"),
+					givenUnitStatus("example-foo@2.service", "2", "172.17.8.102", "9ebb53b04b0d46fb94b4fd1b3f562d2b", "loaded", "loaded"),
+					givenUnitStatus("example-bar@2.service", "2", "172.17.8.102", "9ebb53b04b0d46fb94b4fd1b3f562d2b", "loaded", "loaded"),
+					givenUnitStatus("example-foo@3.service", "3", "172.17.8.103", "e3cb5f13a9164ba5b7eff6c920475e61", "loaded", "loaded"),
+					givenUnitStatus("example-bar@3.service", "3", "172.17.8.103", "e3cb5f13a9164ba5b7eff6c920475e61", "loaded", "loaded"),
+				},
+				Expand: true,
+			},
+			Expected: []string{
+				"Group | Units | FDState | FCState | SAState | Hash | IP | Machine",
+				"",
+				"example@1 | example-foo@1.service | loaded | loaded | inactive | 4311 | 172.17.8.101 | 505e0d7802d7439a924c269b76f34b5f",
+				"example@1 | example-bar@1.service | loaded | loaded | inactive | 4311 | 172.17.8.101 | 505e0d7802d7439a924c269b76f34b5f",
+				"example@2 | example-foo@2.service | loaded | loaded | inactive | 4311 | 172.17.8.102 | 9ebb53b04b0d46fb94b4fd1b3f562d2b",
+				"example@2 | example-bar@2.service | loaded | loaded | inactive | 4311 | 172.17.8.102 | 9ebb53b04b0d46fb94b4fd1b3f562d2b",
+				"example@3 | example-foo@3.service | loaded | loaded | inactive | 4311 | 172.17.8.103 | e3cb5f13a9164ba5b7eff6c920475e61",
+				"example@3 | example-bar@3.service | loaded | loaded | inactive | 4311 | 172.17.8.103 | e3cb5f13a9164ba5b7eff6c920475e61",
+				"",
+			},
+		},
+		// 1 slice of 2 groups with 2 units - only show example group
+		testCase{
+			Comment: "1 slice of 2 groups with 2 units - only show example group",
+			Input: input{
+				Group: "example",
+				USL: UnitStatusList{
+					givenUnitStatus("example-foo@1.service", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "loaded"),
+					givenUnitStatus("example-bar@1.service", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "loaded"),
+					givenUnitStatus("myapp-foo@1.service", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "loaded"),
+					givenUnitStatus("myapp-bar@1.service", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "loaded"),
+				},
+				Expand: false,
+			},
+			Expected: []string{
+				"Group | Units | FDState | FCState | SAState | IP | Machine",
+				"",
+				"example@1 | * | loaded | loaded | inactive | 172.17.8.101 | 505e0d7802d7439a924c269b76f34b5f",
+				"",
+			},
+		},
+		// One group contains different statuses => expand view
+		testCase{
+			Comment: "One group contains different statuses => expand view",
+			Input: input{
+				Group: "example",
+				USL: UnitStatusList{
+					givenUnitStatus("example-foo@1.service", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "loaded", "launched"),
+					givenUnitStatus("example-bar@1.service", "1", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "launched", "launched"),
+					givenUnitStatus("example-foo@2.service", "2", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "launched", "launched"),
+					givenUnitStatus("example-bar@2.service", "2", "172.17.8.101", "505e0d7802d7439a924c269b76f34b5f", "launched", "launched"),
+				},
+				Expand: false,
+			},
+			Expected: []string{
+				"Group | Units | FDState | FCState | SAState | IP | Machine",
+				"",
+				"example@1 | example-foo@1.service | launched | loaded | inactive | 172.17.8.101 | 505e0d7802d7439a924c269b76f34b5f",
+				"example@1 | example-bar@1.service | launched | launched | inactive | 172.17.8.101 | 505e0d7802d7439a924c269b76f34b5f",
+				"example@2 | * | launched | launched | inactive | 172.17.8.101 | 505e0d7802d7439a924c269b76f34b5f",
+				"",
+			},
+		},
+	}
+
+	// execute test cases
+	for _, test := range testCases {
+		got, err := CreateStatus(test.Input.Group, test.Input.USL, test.Input.Expand)
+		Expect(err).To(Not(HaveOccurred()))
+
+		Expect(got).To(Equal(test.Expected), test.Comment)
+	}
+}
+
+func givenUnitStatus(name, sliceID, machineIP, machineID, currentState, desiredState string) fleet.UnitStatus {
+	return fleet.UnitStatus{
+		Current: currentState,
+		Desired: desiredState,
+		Machine: []fleet.MachineStatus{
+			fleet.MachineStatus{
+				ID:            machineID,
+				IP:            net.ParseIP(machineIP),
+				SystemdActive: "inactive",
+				SystemdSub:    "inactive",
+				UnitHash:      "4311",
+			},
+		},
+		Name:    name,
+		SliceID: sliceID,
+	}
+}
